delta_lake/pyproject.toml
[tool.poetry]
name = "delta-lake"
version = "0.1.0"
description = ""
authors = ["Guilherme Samuel <guilherme.silva@condoconta.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
polars = "^1.3.0"
fsspec = "^2024.6.1"
minio = "^7.2.7"
boto3 = "^1.34.152"
python-dotenv = "^1.0.1"
jupyter = "^1.0.0"
notebook = "^7.2.1"
deltalake = "^0.18.2"
pyspark = "^3.5.1"
delta-spark = "^3.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
